backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
transformers==4.35.2
torch==2.1.1
numpy<2.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.32.5

# Note: torch is large (~192MB), if timeout occurs:
# Install torch separately: pip install torch==2.1.1 --timeout 300
# NumPy < 2.0.0 required for torch 2.1.1 compatibility


